Fix pattern not ending in slash not matching directory (#36)
I've found that when having an entry like `/foo/bar @foo` in the CODEOWNERS
file the library properly reports `@foo` as the owner for `/foo/bar`
and `/foo/bar/quux` but not for `/foo/bar/` (note the slash at the
end). This PR adds tests for this case and fixes the regular
expression to take this case into account.